requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
anyascii==0.3.2
    # via wagtail
asgiref==3.8.1
    # via django
beautifulsoup4==4.11.2
    # via wagtail
boto3==1.38.3
    # via -r requirements.in
botocore==1.38.8
    # via
    #   boto3
    #   s3transfer
certifi==2022.12.7
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.1.0
    # via requests
defusedxml==0.7.1
    # via willow
dj-database-url==2.3.0
    # via -r requirements.in
django==5.1.8
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-anymail
    #   django-debug-toolbar
    #   django-filter
    #   django-modelcluster
    #   django-permissionedforms
    #   django-recaptcha
    #   django-storage-url
    #   django-storages
    #   django-stubs-ext
    #   django-taggit
    #   django-tasks
    #   django-treebeard
    #   djangorestframework
    #   laces
    #   wagtail
django-anymail[mailgun]==13.0
    # via -r requirements.in
django-debug-toolbar==5.1.0
    # via -r requirements.in
django-filter==23.4
    # via wagtail
django-modelcluster==6.2.1
    # via wagtail
django-permissionedforms==0.1
    # via wagtail
django-recaptcha==4.1.0
    # via
    #   -r requirements.in
    #   wagtail-django-recaptcha
django-storage-url==0.12.1
    # via -r requirements.in
django-storages==1.14.6
    # via
    #   -r requirements.in
    #   django-storage-url
django-stubs-ext==5.1.3
    # via django-tasks
django-taggit==5.0.1
    # via wagtail
django-tasks==0.6.1
    # via wagtail
django-treebeard==4.7
    # via wagtail
djangorestframework==3.15.1
    # via wagtail
draftjs-exporter==2.1.7
    # via wagtail
et-xmlfile==1.1.0
    # via openpyxl
filetype==1.2.0
    # via willow
fontawesomefree==6.6.0
    # via -r requirements.in
furl==2.1.3
    # via django-storage-url
gevent==24.2.1
    # via -r requirements.in
greenlet==3.2.1
    # via gevent
gunicorn==22.0.0
    # via -r requirements.in
idna==3.4
    # via requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
laces==0.1.1
    # via wagtail
openpyxl==3.1.2
    # via wagtail
orderedmultidict==1.0.1
    # via furl
packaging==25.0
    # via gunicorn
pillow==11.2.1
    # via
    #   -r requirements.in
    #   pillow-heif
    #   wagtail
pillow-heif==0.13.1
    # via willow
psycopg==3.2.6
    # via -r requirements.in
python-dateutil==2.8.2
    # via botocore
pytz==2023.3
    # via django-modelcluster
requests==2.28.2
    # via
    #   django-anymail
    #   wagtail
s3transfer==0.12.0
    # via boto3
sentry-sdk==2.27.0
    # via -r requirements.in
six==1.16.0
    # via
    #   furl
    #   orderedmultidict
    #   python-dateutil
soupsieve==2.4.1
    # via beautifulsoup4
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
telepath==0.3.1
    # via wagtail
typing-extensions==4.12.2
    # via
    #   dj-database-url
    #   django-stubs-ext
    #   django-tasks
    #   psycopg
urllib3==1.26.15
    # via
    #   botocore
    #   django-anymail
    #   requests
    #   sentry-sdk
wagtail==6.4.1
    # via -r requirements.in
wagtail-django-recaptcha==2.1.1
    # via -r requirements.in
whitenoise==6.9.0
    # via -r requirements.in
willow[heif]==1.8.0
    # via
    #   wagtail
    #   willow
zope-event==5.0
    # via gevent
zope-interface==7.2
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
